requirements.txt
# Databricks SDK
databricks-sdk>=0.18.0

# SQL Parsing and Transformation
sqlparse>=0.4.4
sqlglot>=20.0.0

# Data Processing
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging
colorlog>=6.7.0

# CLI
click>=8.1.7
rich>=13.0.0
